// src/Service/Interfaces/IUserData.ts
export interface IUserData{
        login: string;
        name: string;
        email: string;
        celphone: string;
        cgc: string;
        rg: string;
        issuer: string;
        birthday: Date;
        gender: number;
        maritalStatus: number;
        physicalDisability: number;
        disabilityAdditionalDescription: string;
}